parser update and debug
