+++ .idea/dictionaries/SAGE_TP_E480_1.xml
@@ -1,6 +1,8 @@
 <component name="ProjectDictionaryState">
   <dictionary name="SAGE TP-E480-1">
     <words>
+      <w>blockquote</w>
+      <w>emoji</w>
       <w>lihkg</w>
     </words>
   </dictionary>
